Add granular zoom (0.5x steps) and clean up project files
Change font_scale from int to float for finer zoom control (1.0x-8.0x
in 0.5 increments). Update bitmap renderer to handle fractional scales.
Remove stray binaries and temp files, organize .gitignore.

+++ include/render.h
@@ -31,7 +31,7 @@ typedef struct
     Theme         theme;
     size_t        scroll_y;
     size_t        scroll_x;
-    int           font_scale;
+    float         font_scale;
     bool          syntax_enabled;
 } Renderer;
